写完latch 和 两端锁

+++ src/execution/executor_delete.h
@@ -78,6 +78,10 @@ class DeleteExecutor : public AbstractExecutor {
                 }
             }
             if (cond_flag) {
+                if(context_->lock_mgr_->lock_exclusive_on_record(context_->txn_, rid, fh_->GetFd())==false){
+                    throw TransactionAbortException(context_->txn_.get_transaction_id(),AbortReason::GET_LOCK_FAILED);
+                }
+
                 fh_->delete_record(rid, context_);
 
                 for (size_t i = 0; i < indexes.size(); ++i) {
